# refactor: Change simulator state detection logic to use faster APIs