--- a/lib/simulator-xcode-6.js
+++ b/lib/simulator-xcode-6.js
@@ -364,8 +364,29 @@ class SimulatorXcode6 extends EventEmitter {
    * @return {boolean} True if the current Simulator is running.
    */
   async isRunning () {
-    let stat = await this.stat();
-    return stat.state === 'Booted';
+    try {
+      await this.simctl.getEnv('dummy');
+      return true;
+    } catch (e) {
+      return false;
+    }
+  }
+
+  /**
+   * Checks if the simulator is in shutdown state.
+   * This method is necessary, because Simulator might also be
+   * in the transitional Shutting Down state right after the `shutdown`
+   * command has been issued.
+   *
+   * @return {boolean} True if the current Simulator is shut down.
+   */
+  async isShutdown () {
+    try {
+      await this.simctl.getEnv('dummy');
+      return false;
+    } catch (e) {
+      return _.includes(e.stderr, 'Current state: Shutdown');
+    }
   }
 
   /**
@@ -466,8 +487,8 @@ class SimulatorXcode6 extends EventEmitter {
     ];
 
     if (opts.scaleFactor) {
-      const stat = await this.stat();
-      const formattedDeviceName = stat.name.replace(/\s+/g, '-');
+      const {name} = await this.stat();
+      const formattedDeviceName = name.replace(/\s+/g, '-');
       const argumentName = `-SimulatorWindowLastScale-com.apple.CoreSimulator.SimDeviceType.${formattedDeviceName}`;
       args.push(argumentName, opts.scaleFactor);
     }
@@ -524,8 +545,7 @@ class SimulatorXcode6 extends EventEmitter {
     opts = Object.assign({
       startupTimeout: this.startupTimeout,
     }, opts);
-    const {state} = await this.stat();
-    const isServerRunning = state === 'Booted';
+    const isServerRunning = await this.isRunning();
     const isUIClientRunning = await this.isUIClientRunning();
     if (isServerRunning && isUIClientRunning) {
       log.info(`Both Simulator with UDID ${this.udid} and the UI client are currently running`);
@@ -1202,8 +1222,7 @@ class SimulatorXcode6 extends EventEmitter {
     if (_.isString(excludePatterns)) {
       excludePatterns = excludePatterns.split(',').map((x) => x.trim());
     }
-    const {state} = await this.stat();
-    const isServerRunning = state === 'Booted';
+    const isServerRunning = await this.isRunning();
     let plistPath;
     if (isServerRunning) {
       plistPath = path.resolve(await this.getLaunchDaemonsRoot(), 'com.apple.securityd.plist');