--- a/lib/simulator-xcode-9.js
+++ b/lib/simulator-xcode-9.js
@@ -86,20 +86,9 @@ class SimulatorXcode9 extends SimulatorXcode8 {
     if (!_.isEmpty(opts.devicePreferences) || !_.isEmpty(commonPreferences)) {
       await this.updatePreferences(opts.devicePreferences, commonPreferences);
     }
-    const waitForShutdown = async (waitMs = SIMULATOR_SHUTDOWN_TIMEOUT) => {
-      try {
-        await waitForCondition(async () => {
-          const {state} = await this.stat();
-          return state === 'Shutdown';
-        }, {waitMs, intervalMs: 500});
-      } catch (err) {
-        throw new Error(`Simulator is not in 'Shutdown' state after ${waitMs}ms`);
-      }
-    };
     const timer = new timing.Timer().start();
     const shouldWaitForBoot = await startupLock.acquire(this.uiClientBundleId, async () => {
-      const {state: serverState} = await this.stat();
-      const isServerRunning = serverState === 'Booted';
+      const isServerRunning = await this.isRunning();
       const uiClientPid = await this.getUIClientPid();
       if (opts.isHeadless) {
         if (isServerRunning && !uiClientPid) {
@@ -111,12 +100,13 @@ class SimulatorXcode9 extends SimulatorXcode8 {
         }
         try {
           // Stopping the UI client kills all running servers for some early XCode versions. This is a known bug
-          await waitForShutdown(3000);
+          await waitForCondition(async () => await this.isShutdown(), {
+            waitMs: 5000,
+            intervalMs: 100,
+          });
         } catch (e) {
-          const {state} = await this.stat();
-          if (state !== 'Booted') {
-            throw new Error(`Simulator with UDID '${this.udid}' cannot be transitioned to headless mode. ` +
-              `The recent state is '${state}'`);
+          if (!await this.isRunning()) {
+            throw new Error(`Simulator with UDID '${this.udid}' cannot be transitioned to headless mode`);
           }
           return false;
         }
@@ -127,16 +117,10 @@ class SimulatorXcode9 extends SimulatorXcode8 {
           log.info(`Both Simulator with UDID '${this.udid}' and the UI client are currently running`);
           return false;
         }
-        if (!['Shutdown', 'Booted'].includes(serverState)) {
-          if (serverState !== 'Shutting Down') {
-            log.info(`Simulator '${this.udid}' is in '${serverState}' state. Trying to shutdown...`);
-            try {
-              await this.shutdown();
-            } catch (err) {
-              log.warn(`Error on Simulator shutdown: ${err.message}`);
-            }
-          }
-          await waitForShutdown();
+        if (isServerRunning) {
+          log.info(`Simulator '${this.udid}' is booted while its UI is not visible. ` +
+            `Trying to restart it with the Simulator window visible`);
+          await this.shutdown({timeout: SIMULATOR_SHUTDOWN_TIMEOUT});
         }
         await this.launchWindow(uiClientPid, opts);
       }
@@ -166,9 +150,8 @@ class SimulatorXcode9 extends SimulatorXcode8 {
    * Boots simulator if not already booted.
    */
   async boot () {
-    const {state} = await this.stat();
-    if (['Booted', 'Booting'].includes(state)) {
-      log.info(`Simulator '${this.udid}' is already in ${state} state`);
+    if (await this.isRunning()) {
+      log.info(`Simulator '${this.udid}' is already running`);
       return;
     }
 
@@ -284,15 +267,37 @@ class SimulatorXcode9 extends SimulatorXcode8 {
   }
 
   /**
+   * @typedef {Object} ShutdownOptions
+   * @property {?number|string} timeout The number of milliseconds to wait until
+   * Simulator is shut down completely. No wait happens if the timeout value is not set
+   */
+
+  /**
    * Shut down the current Simulator.
    * @override
+   *
+   * @param {?ShutdownOptions} opts
+   * @throws {Error} If Simulator fails to transition into Shutdown state after
+   * the given timeout
    */
-  async shutdown () {
-    const {state} = await this.stat();
-    if (state === 'Shutdown') {
-      return;
+  async shutdown (opts = {}) {
+    if (await this.isRunning()) {
+      await retryInterval(5, 500, this.simctl.shutdownDevice.bind(this.simctl));
+      const {
+        timeout,
+      } = opts;
+      const waitMs = parseInt(timeout, 10);
+      if (waitMs > 0) {
+        try {
+          await waitForCondition(async () => await this.isShutdown(), {
+            waitMs,
+            intervalMs: 100,
+          });
+        } catch (err) {
+          throw new Error(`Simulator is not in 'Shutdown' state after ${waitMs}ms`);
+        }
+      }
     }
-    await retryInterval(5, 500, this.simctl.shutdownDevice.bind(this.simctl));
   }
 
   /**