--- a/package.json
+++ b/package.json
@@ -37,7 +37,7 @@
     "bluebird": "^3.5.1",
     "fkill": "^7.0.0",
     "lodash": "^4.2.1",
-    "node-simctl": "^6.1.0",
+    "node-simctl": "^6.3.0",
     "openssl-wrapper": "^0.3.4",
     "semver": "^7.0.0",
     "source-map-support": "^0.5.3",