The `stat` method internally invokes `simctl list`, which could be pretty slow, especially if multiple simulators are installed on the system. I've switched the state detection API to use `getenv` call as a workaround. In opposite to the `list` call it only queries single simulator, which makes the overall performance better. On my machine the timings were:

```
$ time xcrun simctl getenv 0B2C9CE6-29AC-4494-B64F-CE711619146E dummy
...
xcrun simctl getenv 0B2C9CE6-29AC-4494-B64F-CE711619146E dummy  0.08s user 0.09s system 88% cpu 0.184 total
```

```
$ % time xcrun simctl list
....
xcrun simctl list  0.11s user 0.20s system 43% cpu 0.723 total
```